requirements.txt
psycopg2-binary>=2.9.0
python-dotenv>=0.19.0
jsonschema>=4.0.0
requests>=2.28.0
